Homework-7
