+++ constants/index.js
@@ -1,0 +1,7 @@
+export const HIT = {
+    head: 30,
+    body: 25,
+    foot: 20,
+};
+
+export const ATTACK = ['head', 'body', 'foot'];
