--- func.js
@@ -1,5 +1,0 @@
-function getRandom(number) {
-    return Math.ceil(Math.random() * number);
-}
-export default getRandom;
-
